CareerPath-AI/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.8.1",
    "pdfplumber>=0.11.8",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "reportlab>=4.4.5",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "streamlit-extras>=0.7.8",
    "streamlit-lottie>=0.0.5",
]
